+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Enabled pagination via itemsPerPage parameter
 - Introduced AaaS security roles (ROLE_AAAS_USER, ROLE_AAAS_ADMIN)
+- We can add users via a console command now (php bin/console acl:create-user EMAIL PASSWORD [--admin])
 
 ### Changed